This commit consists of a major change in approach: Instead of estimating an upper-bound of the 'ln' function in order to predict the highest precision which is legitimate for the given input, we call the 'ln' function with the highest precision possible, and then check how much we need to shift the result in order to make it legitimate for function 'fixedExp'.
This method allows us to find not only a good precision for any given input, but the highest possible precision for any given input.
It therefore also allows us to define MIN_PRECISION as low as we want, without the risk of using a non-optimal precision for some given input.

Methematically, it relies on the following fact: fixedLog2(x,precision+1) >= fixedLog2(x,precision)<<1.

In terms of coding, this method is much simpler, since it consists of a single flow: Sale/Purchase --> power --> ln, findPositionInMaxExpArray, fixedExp.

In terms of accuracy, it achieves the best possible results; any remaining inaccuracy is due to the finite number of Taylor coefficients (34).

In terms of performance, here are some empiric measurements:
- Without optimization
  - Sale: At most ~6,500 gas worse
  - Purchase: At least ~2,500 gas better
- With optimization
  - Sale: At most ~2,000 gas worse
  - Purchase: At least ~2,000 gas better
Of course, in the worst case, since the previous method would apply the highest precision possible, both methods are expected to yield more or less the same performance.